refactor(desktop): extract the beta email signup, onboarding toggle card, shortcut pill, and toast bodies shared by two surfaces each
Five frontend copy-paste pairs from the `jscpd-frontend` inventory, each replaced by one shared module so the two surfaces can no longer drift apart:

- `StepBeta.svelte` and `UpdatesSection.svelte` both carried the beta contact email field's state and handlers (persist on keystroke, subscribe on commit, typed feedback). Now `settings/sections/beta-email-signup.svelte.ts` (`createBetaEmailSignup()`, per-mount `$state` behind getters).
- `StepBeta.svelte` and `StepOptional.svelte` shared the toggle card's markup and styles. Now `onboarding/OnboardingToggleCard.svelte`.
- `GlobalShortcutRow.svelte` and `KeyboardShortcutsSection.svelte` shared the editable shortcut pill's styles. Now `settings/sections/ShortcutPill.svelte` (class names preserved for the existing tests).
- `OperationFailedToastContent.svelte` and `OperationFailuresToastContent.svelte` shared the toast body. Now `status-corner/OperationFailureToastBody.svelte`.
- `AutoSendToastContent.svelte` and `ErrorReportToastContent.svelte` shared the sent-report toast body. Now `error-reporter/SentReportToastBody.svelte`.

`GitSection` and `SearchSection` keep their 18 shared lines: that is the settings-section registry boilerplate every section carries, and hiding it would obscure the contract.

Behavior preserved as-is, including the `analytics.email` listener that both surfaces subscribe at init without unsubscribing. Colocated `CLAUDE.md` / `DETAILS.md` module maps updated.

Verified: svelte-check, stylelint, css-unused green; eslint green on the touched files; targeted vitest run of the touched dirs (41 files, 487 tests) green.
